New tool: thc_livecheck.py
Simple python script that will take the output from thc_recon.py and will find which hosts are publicly resolvable, while saving the hosts that resolve to an internal ip to a seperate file. The tool can be used to find which public host has a webserver on the most common ports, attempts to fingerprint the server, saves the title of the root page and takes a screen shot.